+++ pulsar/utils/structures.py
@@ -8,7 +8,9 @@ MultiValueDict
    :members:
    :member-order: bysource
    
-   
+
+.. _attribute-dictionary:
+
 AttributeDictionary
 ~~~~~~~~~~~~~~~~~~~~~~~